Fantail Refactoring
