+++ Java/OpenmlWebapplication/src/org/openml/webapplication/features/AttributeMetafeatures.java
@@ -1,0 +1,7 @@
+package org.openml.webapplication.features;
+
+/**
+ * @author Jakub Smid <jakub.smid@merck.com>
+ */
+public class AttributeMetafeatures {
+}
