+++ Java/OpenmlWebapplication/src/org/openml/webapplication/features/FantailConnector.java
@@ -18,15 +18,8 @@
  *  
  */
 package org.openml.webapplication.features;
-
-import java.io.FileReader;
-import java.util.*;
-
-import org.apache.commons.lang3.ArrayUtils;
-import org.json.JSONArray;
-import org.json.JSONException;
+import com.thoughtworks.xstream.XStream;
 import org.openml.apiconnector.algorithms.Conversion;
-import org.openml.apiconnector.algorithms.QueryUtils;
 import org.openml.apiconnector.io.OpenmlConnector;
 import org.openml.apiconnector.settings.Constants;
 import org.openml.apiconnector.xml.DataQuality;
@@ -36,21 +29,13 @@ import org.openml.apiconnector.xml.DataSetDescription;
 import org.openml.apiconnector.xstream.XstreamXmlMapping;
 import org.openml.webapplication.fantail.dc.Characterizer;
 import org.openml.webapplication.fantail.dc.StreamCharacterizer;
-import org.openml.webapplication.fantail.dc.landmarking.GenericLandmarker;
-import org.openml.webapplication.fantail.dc.statistical.AttributeEntropy;
-import org.openml.webapplication.fantail.dc.statistical.NominalAttDistinctValues;
-import org.openml.webapplication.fantail.dc.statistical.SimpleMetaFeatures;
-import org.openml.webapplication.fantail.dc.statistical.Statistical;
-import org.openml.webapplication.fantail.dc.stream.ChangeDetectors;
-import org.openml.webapplication.fantail.dc.stream.RunChangeDetectorTask;
-
-import com.thoughtworks.xstream.XStream;
-
 import weka.core.Instances;
 import weka.core.OptionHandler;
 import weka.core.Utils;
 import weka.filters.Filter;
 import weka.filters.unsupervised.attribute.StringToNominal;
+
+import java.util.*;
 
 public class FantailConnector {
     public class QualityResult{
@@ -94,155 +79,50 @@ public class FantailConnector {
     }
 
 	private final Integer window_size;
-
-	private final String preprocessingPrefix = "-E \"weka.attributeSelection.CfsSubsetEval -P 1 -E 1\" -S \"weka.attributeSelection.BestFirst -D 1 -N 5\" -W ";
-	private final String cp1NN = "weka.classifiers.lazy.IBk";
-	private final String cpNB = "weka.classifiers.bayes.NaiveBayes";
-	private final String cpASC = "weka.classifiers.meta.AttributeSelectedClassifier";
-	private final String cpDS = "weka.classifiers.trees.DecisionStump";
 	
 	private static final XStream xstream = XstreamXmlMapping.getInstance();
-	private ArrayList<CharacterizerWrapper> batchCharacterizers = new ArrayList<>();
-	
-	private static StreamCharacterizer[] streamCharacterizers;
-	private static OpenmlConnector apiconnector;
+	private OpenmlConnector apiconnector;
+	private DatabaseUtils dbUtils;
+	private GlobalMetafeatures globalMetafeatures;
+
 	
 	public FantailConnector(OpenmlConnector ac, Integer dataset_id, boolean random, String priorityTag, Integer interval_size) throws Exception {
-		int expectedQualities = 0; 
-		apiconnector = ac;
 		window_size = interval_size;
+		apiconnector = ac;
+		dbUtils = new DatabaseUtils(apiconnector);
+        globalMetafeatures = new GlobalMetafeatures(window_size);
 
-		Characterizer[] characterizers = {
-                new SimpleMetaFeatures(), // done before, but necessary for streams
-                new Statistical(),
-                new NominalAttDistinctValues(),
-                new AttributeEntropy(),
-                new GenericLandmarker("kNN1N", cp1NN, 2, null),
-                new GenericLandmarker("NaiveBayes", cpNB, 2, null),
-                new GenericLandmarker("DecisionStump", cpDS, 2, null),
-                new GenericLandmarker("CfsSubsetEval_kNN1N", cpASC, 2, Utils.splitOptions(preprocessingPrefix + cp1NN)),
-                new GenericLandmarker("CfsSubsetEval_NaiveBayes", cpASC, 2, Utils.splitOptions(preprocessingPrefix + cpNB)),
-                new GenericLandmarker("CfsSubsetEval_DecisionStump", cpASC, 2, Utils.splitOptions(preprocessingPrefix + cpDS))
-        };
-		// additional parameterized batch landmarkers
-		String zeros = "0";
-		for( int i = 1; i <= 3; ++i ) {
-			zeros += "0";
-			String[] j48Option = { "-C", "." + zeros + "1" };
-			characterizers = ArrayUtils.add(characterizers, new GenericLandmarker("J48." + zeros + "1.", "weka.classifiers.trees.J48", 2, j48Option));
-			
-			String[] repOption = { "-L", "" + i };
-            characterizers = ArrayUtils.add(characterizers, new GenericLandmarker("REPTreeDepth" + i, "weka.classifiers.trees.REPTree", 2, repOption));
-			
-			String[] randomtreeOption = { "-depth", "" + i };
-            characterizers = ArrayUtils.add(characterizers, new GenericLandmarker("RandomTreeDepth" + i, "weka.classifiers.trees.RandomTree", 2, randomtreeOption));
-		}
-		for (Characterizer characterizer: characterizers){
-		    batchCharacterizers.add(new CharacterizerWrapper(characterizer));
-        }
-		for(CharacterizerWrapper characterizer : batchCharacterizers) {
-			expectedQualities += characterizer.getNumMetaFeatures();
-		}
-		// add expected qualities for stream landmarkers (initialized later)
-		expectedQualities += RunChangeDetectorTask.getNumMetaFeatures();
-		
-		
 		if(dataset_id != null) {
 			Conversion.log("OK", "Process Dataset", "Processing dataset " + dataset_id + " on special request. ");
-			extractFeatures(dataset_id, window_size);
+			computeMetafeatures(dataset_id);
+
 		} else {
-			dataset_id = getDatasetId(expectedQualities, window_size, random, priorityTag);
+
+			dataset_id = dbUtils.getDatasetId(getExpectedNumberOfMetafeatures(), window_size, random, priorityTag);
 			while( dataset_id != null ) {
 				Conversion.log("OK", "Process Dataset", "Processing dataset " + dataset_id + " as obtained from database. ");
-				extractFeatures(dataset_id, window_size);
-				dataset_id = getDatasetId(expectedQualities, window_size, random, priorityTag);
+				computeMetafeatures(dataset_id);
+				dataset_id = dbUtils.getDatasetId(getExpectedNumberOfMetafeatures(), window_size, random, priorityTag);
 			}
 			Conversion.log("OK", "Process Dataset", "No more datasets to process. ");
 		}
 	}
-	
-	public Integer getDatasetId(int expectedQualities, Integer window_size, boolean random, String priorityTag) throws JSONException, Exception {
-		String tagJoin = "";
-		String tagSelect = "";
-		String tagSort = "";
-		if (priorityTag != null) {
-			tagSelect = ", t.tag ";
-			tagSort = "t.tag DESC, "; // to avoid NULL values first
-			tagJoin = "LEFT JOIN dataset_tag t ON q.data = t.id AND t.tag = '" + priorityTag + "' ";
-		}
-		
-		String sql = 
-			"SELECT `d`.`did`, `q`.`value` AS `numInstances`, `i`.`interval_end` - `i`.`interval_start` AS `interval_size`, " +
-			"CEIL(`q`.`value` / " + window_size + ") AS `numIntervals`, " +
-			"(COUNT(*) / CEIL(`q`.`value` / " + window_size + ")) AS `qualitiesPerInterval`, " +
-			"COUNT(*) AS `qualities` " + tagSelect + 
-			"FROM `data_quality` `q` " + tagJoin + 
-			", `dataset` `d`" +
-			"LEFT JOIN `data_quality_interval` `i` ON `d`.`did` = `i`.`data` AND `i`.`interval_end` - `i`.`interval_start` =  " + window_size + " " +
-			"WHERE `q`.`quality` IS NOT NULL " +
-			"AND `d`.`did` = `q`.`data` " +
-			"AND `q`.`quality` = 'NumberOfInstances'  " +
-			"AND `d`.`error` = 'false' AND `d`.`processed` IS NOT NULL " +
-			"GROUP BY `d`.`did` " +
-			"HAVING (COUNT(*) / CEIL(`q`.`value` / " + window_size + ")) < " + expectedQualities + " " +
-			"ORDER BY " + tagSort + "`qualitiesPerInterval` ASC LIMIT 0,100; ";
-		
-		if(window_size == null) {
-			sql = 
-				"SELECT q.data, COUNT(*) AS `numQualities`" + tagSelect + 
-				" FROM data_quality q " + tagJoin + 
-				" GROUP BY q.data HAVING numQualities BETWEEN 0 AND " + (expectedQualities-1) + 
-				" ORDER BY " + tagSort + " q.data LIMIT 0,100";
-		}
-			
-		Conversion.log("OK", "FantailQuery", sql);
-		JSONArray runJson = (JSONArray) apiconnector.freeQuery(sql).get("data");
-		
-		
-		int randomint = 0;
-		
-		if (random) {
-			Random randomgen = new Random(System.currentTimeMillis());
-			randomint = Math.abs(randomgen.nextInt());
-		}
-		
-		if(runJson.length() > 0) {
-			int dataset_id = ((JSONArray) runJson.get(randomint % runJson.length())).getInt(0);
-			return dataset_id;
-		} else {
-			return null;
-		}
+
+	private int getExpectedNumberOfMetafeatures(){
+		//TODO: modify code to work with expected attribute metafeatures,
+		//TODO: perhaps modify it so it really verifies that all valid qualities are computed
+		return globalMetafeatures.getExpectedQualities();
 	}
+
+	private void computeMetafeatures(int datasetId) throws Exception {
+	    DataSetDescription dsd = dbUtils.GetDatasetDescription(datasetId);
+	    Instances dataset = dbUtils.getDataset(dsd);
+	    List<String> qualitiesAvailable = dbUtils.GetQualitiesAvailable(datasetId, window_size);
+        extractFeatures(dsd, dataset, qualitiesAvailable);
+    }
 	
-	private boolean extractFeatures(Integer did, Integer interval_size) throws Exception {
-		Conversion.log("OK", "Extract Features", "Start extracting features for dataset: " + did);
-		
-		List<String> qualitiesAvailable = Arrays.asList(apiconnector.dataQualities(did).getQualityNames());
-		if (interval_size != null) {
-			// alternative approach to knowing which features are already complete. 
-			String sql = 
-				"SELECT i.quality, CEIL(`q`.`value` / " + window_size + ") AS `numIntervals`, COUNT(*) AS `present` " + 
-				"FROM `data_quality` `q`, `dataset` `d` LEFT JOIN `data_quality_interval` `i` ON `d`.`did` = `i`.`data` AND `i`.`interval_end` - `i`.`interval_start` =  " + window_size + " " +
-				"WHERE `d`.`did` = `q`.`data` AND `q`.`quality` = 'NumberOfInstances'  AND `d`.`error` = 'false' AND `d`.`processed` IS NOT NULL AND d.did = " + did + " " +
-				"GROUP BY `d`.`did`,`i`.`quality` HAVING `present` = `numIntervals`";
-			Conversion.log("OK", "FantailQuery for interval queries", sql);
-			qualitiesAvailable = Arrays.asList(QueryUtils.getStringsFromDatabase(apiconnector, sql));
-		}
-		
-		
-		// TODO: initialize this properly!!!!!!
-		streamCharacterizers = new StreamCharacterizer[1]; 
-		streamCharacterizers[0] = new ChangeDetectors(interval_size);
-		
-		DataSetDescription dsd = apiconnector.dataGet(did);
-		
-		Conversion.log("OK", "Extract Features", "Start downloading dataset: " + did);
-		
-		Instances dataset = new Instances(new FileReader(dsd.getDataset(apiconnector.getApiKey())));
-		
-		if (dsd.getDefault_target_attribute() == null) {
-			throw new RuntimeException("Default target attribute is null. ");
-		}
+	private boolean extractFeatures(DataSetDescription dsd, Instances dataset, List<String> qualitiesAvailable) throws Exception {
+		Conversion.log("OK", "Extract Features", "Start extracting features for dataset: " + dsd.getId());
 		
 		dataset.setClass(dataset.attribute(dsd.getDefault_target_attribute()));
 		if (dsd.getRow_id_attribute() != null) {
@@ -259,7 +139,7 @@ public class FantailConnector {
 		}
 
 		// first run stream characterizers
-		for(StreamCharacterizer sc : streamCharacterizers) {
+		for(StreamCharacterizer sc : globalMetafeatures.getStreamCharacterizers()) {
 
 			if (qualitiesAvailable.containsAll(Arrays.asList(sc.getIDs())) == false) {
 				Conversion.log("OK", "Extract Features", "Running Stream Characterizers (full data)");
@@ -270,19 +150,19 @@ public class FantailConnector {
 		}
 		
 		List<Quality> qualities = new ArrayList<DataQuality.Quality>();
-		if (interval_size != null) {
+		if (window_size != null) {
 			Conversion.log("OK", "Extract Features", "Running Batch Characterizers (partial data)");
 			
-			for (int i = 0; i < dataset.numInstances(); i += interval_size) {
+			for (int i = 0; i < dataset.numInstances(); i += window_size) {
 				if (apiconnector.getVerboselevel() >= Constants.VERBOSE_LEVEL_ARFF) {
-					Conversion.log("OK", "FantailConnector", "Starting window [" + i + "," + (i + interval_size) + "> (did = " + did + ",total size = " + dataset.numInstances() + ")");
+					Conversion.log("OK", "FantailConnector", "Starting window [" + i + "," + (i + window_size) + "> (did = " + dsd.getId() + ",total size = " + dataset.numInstances() + ")");
 				}
-				qualities.addAll(datasetCharacteristics(dataset, i, interval_size, null));
+				qualities.addAll(datasetCharacteristics(dataset, i, window_size, null));
 				
-				for(StreamCharacterizer sc : streamCharacterizers) {
+				for(StreamCharacterizer sc : globalMetafeatures.getStreamCharacterizers()) {
 					// preventing nullpointer exception (if stream characterizer was already run)
 					if (qualitiesAvailable.containsAll(Arrays.asList(sc.getIDs())) == false) {
-						qualities.addAll(hashMaptoList(sc.interval(i), i, interval_size));
+						qualities.addAll(hashMaptoList(sc.interval(i), i, window_size));
 					}
 				}
 			}
@@ -290,7 +170,7 @@ public class FantailConnector {
 		} else {
 			Conversion.log("OK", "Extract Features", "Running Batch Characterizers (full data, might take a while)");
 			qualities.addAll(datasetCharacteristics(dataset, null, null, qualitiesAvailable));
-			for (StreamCharacterizer sc : streamCharacterizers) {
+			for (StreamCharacterizer sc : globalMetafeatures.getStreamCharacterizers()) {
 				Map<String, Double> streamqualities = sc.global();
 				if (streamqualities != null) {
 					qualities.addAll(hashMaptoList(streamqualities, null, null));
@@ -299,10 +179,10 @@ public class FantailConnector {
 		}
 		Conversion.log("OK", "Extract Features", "Done generating features, start wrapping up");
 		if(qualities.size()>0){
-			DataQuality dq = new DataQuality(did, qualities.toArray(new Quality[qualities.size()]));
+			DataQuality dq = new DataQuality(dsd.getId(), qualities.toArray(new Quality[qualities.size()]));
 			String strQualities = xstream.toXML(dq);
 			
-			DataQualityUpload dqu = apiconnector.dataQualitiesUpload(Conversion.stringToTempFile(strQualities, "qualities_did_" + did, "xml"));
+			DataQualityUpload dqu = apiconnector.dataQualitiesUpload(Conversion.stringToTempFile(strQualities, "qualities_did_" + dsd.getId(), "xml"));
 			Conversion.log("OK", "Extract Features", "DONE: " + dqu.getDid());
 		}
 		else{
@@ -326,11 +206,11 @@ public class FantailConnector {
 			// todo: use StringToNominal filter? might be too expensive
 		}
 		
-		for(CharacterizerWrapper dc : batchCharacterizers) {
+		for(Characterizer dc : globalMetafeatures.getCharacterizers()) {
 			if (qualitiesAvailable != null && qualitiesAvailable.containsAll(Arrays.asList(dc.getIDs())) == false) { 
 				Conversion.log("OK","Extract Batch Features", dc.getClass().getName() + ": " + Arrays.toString(dc.getIDs()));
-				Map<String,QualityResult> qualities = dc.characterize(intervalData);
-				result.addAll(qualityResultToList(qualities, start, interval_size));
+				Map<String,Double> qualities = dc.characterize(intervalData);
+				result.addAll(hashMaptoList(qualities, start, interval_size));
 			} else {
 				Conversion.log("OK","Extract Batch Features", dc.getClass().getName() + " - already in database");
 			}
